+++ frontend/src/pages/guest/imageUploadPage/ImageUploadPage.tsx
@@ -1,12 +1,10 @@
 import rocketIcon from '@assets/images/rocket.png';
-import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { ReactComponent as ArrowUpSvg } from '../../../@assets/icons/upwardArrow.svg';
 import FloatingActionButton from '../../../components/@common/buttons/floatingActionButton/FloatingActionButton';
 import FloatingIconButton from '../../../components/@common/buttons/floatingIconButton/FloatingIconButton';
 import HighlightText from '../../../components/@common/highlightText/HighlightText';
 import GuestImageGrid from '../../../components/@common/imageLayout/imageGrid/guestImageGrid/GuestImageGrid';
-import InputModal from '../../../components/@common/modal/inputModal/InputModal';
 import PhotoModal from '../../../components/@common/modal/photoModal/PhotoModal';
 import SpaceHeader from '../../../components/header/spaceHeader/SpaceHeader';
 import LoadingLayout from '../../../components/layout/loadingLayout/LoadingLayout';
@@ -17,6 +15,7 @@ import { useOverlay } from '../../../contexts/OverlayProvider';
 import useFileUpload from '../../../hooks/@common/useFileUpload';
 import useIntersectionObserver from '../../../hooks/@common/useIntersectionObserver';
 import useLeftTimer from '../../../hooks/@common/useLeftTimer';
+import useGuestNickName from '../../../hooks/useGuestNickName';
 import useSpaceCodeFromPath from '../../../hooks/useSpaceCodeFromPath';
 import useSpaceInfo from '../../../hooks/useSpaceInfo';
 import { ScrollableBlurArea } from '../../../styles/@common/ScrollableBlurArea.styles';
@@ -38,67 +37,9 @@ const ImageUploadPage = () => {
   const shouldShowFakeUploadBox = isNoData || isEarlyTime || isSpaceExpired;
 
   const overlay = useOverlay();
-
   const navigate = useNavigate();
 
-  // // TODO : 처음에는 확인만, 그 수정에서는 취소가 보이도록
-  // TODO : 랜덤 닉네임 제공
-
-  const checkNickNameLength = (nickName: string) => {
-    if (nickName.length > 10) {
-      return false;
-    }
-    return true;
-  };
-
-  type NickNameModalMode = 'create' | 'edit';
-  const [nickName, setNickName] = useState('');
-  const showNickNameModal = async (mode: NickNameModalMode) => {
-    const defaultInputModalProps = {
-      description: (
-        <HighlightText
-          text="닉네임을 입력해 주세요"
-          highlightTextArray={['닉네임']}
-          fontStyle="header03"
-          highlightColorStyle="primary"
-        />
-      ),
-      subDescription: '공백 없이 10자까지 입력할 수 있어요.',
-      placeholder: '닉네임을 입력해 주세요',
-      confirmText: '확인',
-    };
-
-    const editModeInputModalProps = {
-      ...defaultInputModalProps,
-      cancelText: '취소',
-    };
-
-    const result = await overlay<{ value: string }>(
-      <InputModal
-        {...(mode === 'edit'
-          ? editModeInputModalProps
-          : defaultInputModalProps)}
-        initialValue=""
-        validation={checkNickNameLength}
-        errorMessage="10자 이하로 입력해주세요."
-      />,
-      {
-        clickOverlayClose: true,
-      },
-    );
-    if (result) {
-      setNickName(result.value);
-    }
-  };
-
-  // const submitNickName = () => {
-  //   // TODO : api 연동
-  // };
-
-  //biome-ignore lint/correctness/useExhaustiveDependencies: 초기 모달 표시
-  useEffect(() => {
-    showNickNameModal('create');
-  }, []);
+  const { nickName, showNickNameModal } = useGuestNickName();
 
   const navigateToUploadComplete = () => {
     navigate(ROUTES.COMPLETE.UPLOAD, {
